02 Pytorch/022.+Implementing+gradio+app+inference+backend(gradio-app-1-chkpt-22)/requirements.txt
gradio
torch
torchvision
opencv-python
pillow